+++ rally/plugins/__init__.py
@@ -32,16 +32,6 @@ def load():
         opts.register()
 
         discover.import_modules_from_package("rally.plugins.common")
-        try:
-            import rally_openstack  # noqa
-        except ImportError:
-            discover.LOG.warning(
-                "OpenStack plugins moved to the separate package "
-                "(see https://pypi.org/project/rally-openstack). In-tree "
-                "OpenStack plugins will be removed from the Rally main package"
-                " soon.")
-            discover.import_modules_from_package("rally.plugins.openstack")
-            discover.import_modules_from_package("rally.plugins.workload")
 
         packages = discover.find_packages_by_entry_point()
         for package in packages: